// src/main/models/entity/credentials/Users.ts
export interface UsersAttributes {
  id: string;
  fullname: string;
  address?: string;
  noTelp?: string;
  email: string;
  passwordDefault: string;
  password: string;
  verifCode: string;
  lastSignIn?: Date | null;
  isOnline?: boolean;
  lang?: string;
  userType?: typeof UsersTypeList[number];
  status?: typeof UsersStatusList[number];
  tokenGcp?: string;
  createdBy: string;
  createdAt: Date;
  updatedBy?: string;
  updatedAt?: Date;
  deletedAt?: Date;
  isSuper?: boolean;
}

type UsersAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type UsersCreationAttributes = Omit<
  UsersAttributes,
  UsersAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'users',
  paranoid: true,
  schema: 'credentials',
  underscored: true,
})
class Users
  extends Model<UsersAttributes | UsersCreationAttributes>
  implements UsersAttributes {
  @Column({
    allowNull: false,
    type: DataType.STRING,
    defaultValue: Sequelize.literal('uuid_generate_v7()'),
    primaryKey: true,
  })
  id!: string;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  fullname!: string;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  address?: string;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  noTelp?: string;


  @Index
  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  email!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
    defaultValue: 'password',
  })
  password!: string;

  @Column({
    allowNull: true,
    type: DataType.STRING,
    defaultValue: 'password',
  })
  passwordDefault!: string;

  @Column({
    allowNull: true,
    type: DataType.STRING,
    defaultValue: 'password',
  })
  verifCode!: string;

  @Column({
    type: DataType.BOOLEAN,
    defaultValue: false,
  })
  isOnline!: boolean;

  @Column({
    type: DataType.BOOLEAN,
    defaultValue: false,
  })
  isSuper!: boolean;


  @Column({
    type: DataType.STRING,
    defaultValue: "id"
  })
  lang?: string;

  @Column({
    type: DataType.DATE,
  })
  lastSignIn!: Date;

  @Column({
    allowNull: false,
    type: DataType.ENUM(...UsersTypeList),
    defaultValue: "PERSONAL"
  })
  userType!: typeof UsersTypeList[number]

  @Column({
    allowNull: false,
    type: DataType.ENUM(...UsersStatusList),
    defaultValue: "NEED VERIFICATION"
  })
  status!: typeof UsersStatusList[number]

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  tokenGcp?: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  updatedBy!: string;

  @Column({
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;

  @HasMany(() => UserRoles, {
    foreignKey: 'userId',
    sourceKey: 'id',
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  roles!: UserRoles[];

  @HasMany(() => UsersAssigment, {
    foreignKey: 'userId',
    sourceKey: 'id',
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  usersAssignment!: UsersAssigment[];

  @HasOne(() => Employee, {
    foreignKey: 'userId',
    sourceKey: 'id',
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  employee!: Employee;

  @HasMany(() => CompanyUser, {
    foreignKey: 'userId',
    sourceKey: 'id',
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  userCompany!: CompanyUser[];

}